+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- Add `#[must_use]` to constructors.
+
 ## [0.2.1] - 2022-03-17
 
 - Implement aarch64 outline-atomics.